Switch runtime state sizing helpers to f32 storage

+++ src/compiler/__tests__/steel-thread.test.ts
@@ -35,7 +35,7 @@ function stateFor(program: ReturnType<typeof compileOk>) {
   const schedule = program.schedule as ScheduleIR;
   const sizes = computeStorageSizes(program.slotMeta);
   return createRuntimeState(
-    sizes.f64,
+    sizes.f32,
     schedule.stateSlotCount,
     0, // eventSlotCount
     0, // eventExprCount
